refactor: combine test folders
js/__test__ and js/tsrc/__test__ have been merged into js/__test__
