+++ Source/Plugins/Core/com.equella.core/js/__tests__/tsrc/mainui/ErrorPage.test.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { createMount } from "@material-ui/core/test-utils";
-import ErrorPage from "../../mainui/ErrorPage";
+import ErrorPage from "../../../tsrc/mainui/ErrorPage";
 
 jest.mock("@material-ui/core/styles", () => ({
   makeStyles: () => () => ({ errorPage: "mock-classname" })
