+++ Source/Plugins/Core/com.equella.core/js/__tests__/tsrc/util/commonstrings.test.ts
@@ -1,4 +1,4 @@
-import { commonString } from "../../util/commonstrings";
+import { commonString } from "../../../tsrc/util/commonstrings";
 
 describe("commonstrings", () => {
   describe("commonString", () => {
