+++ Source/Plugins/Core/com.equella.core/js/__tests__/tsrc/util/dictionary.test.ts
@@ -1,4 +1,4 @@
-import { properties } from "../../util/dictionary";
+import { properties } from "../../../tsrc/util/dictionary";
 
 describe("dictionary", () => {
   describe("properties", () => {
