+++ Source/Plugins/Core/com.equella.core/js/__tests__/tsrc/util/langstrings.test.ts
@@ -3,7 +3,7 @@ import {
   Sizes,
   prepLangStrings,
   initStrings
-} from "../../util/langstrings";
+} from "../../../tsrc/util/langstrings";
 
 describe("langstrings", () => {
   describe("formatSize", () => {
