+++ Source/Plugins/Core/com.equella.core/js/tsconfig.json
@@ -21,7 +21,7 @@
       "~*": ["./entrybuild/*"]
     }
   },
-  "include": ["tsrc/**/*", "__test__/**/*.ts", "__stories__/**/*.tsx"],
+  "include": ["tsrc/**/*", "__test__/**/*", "__stories__/**/*"],
   "exclude": [
     "node_modules",
     "build",
